lab4/bot/requirements.txt
python-telegram-bot=20.6